requirements.txt
flask
cryptography
pycryptodome
pyqrcode
pypng
gunicorn
werkzeug